requirements.txt
asgiref==2.0.1
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.3.0
Django==2.0
django-filter==1.1.0
djangorestframework==3.7.3
django-solo
hyperlink==17.3.1
incremental==17.5.0
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
pypiwin32==220
pytz==2017.3
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
zope.interface==4.4.3
